Move db settings to storage pref fragment
