+++ app/src/main/res/xml/preferences_collection.xml
@@ -30,8 +30,9 @@
         app:useSimpleSummaryProvider="false" />
 
     <PreferenceCategory
-        android:key="transfer_defaults"
+        android:key="category_collection_sounds"
         android:title="@string/preferences_sound_title"
+        app:allowDividerAbove="false"
         app:iconSpaceReserved="false">
 
         <CheckBoxPreference
@@ -57,8 +58,9 @@
     </PreferenceCategory>
 
     <PreferenceCategory
-        android:key="transfer_defaults"
+        android:key="category_appearance_layout"
         android:title="@string/preferences_layout_title"
+        app:allowDividerAbove="false"
         app:iconSpaceReserved="false">
 
         <SeekBarPreference
